[C++, Java] fix alignment issues by making sure that MessagePrefix ist 8 byte aligned
this is achieved by limiting MessagePrefix to a total size of 8 bytes by using

* 2 bytes for the version
* 2 bytes for the signature size
* 4 bytes for the message size

+++ cpp/libsmrf/MessageSerializerImpl.inl
@@ -122,7 +122,7 @@ public:
     {
         const bool isEncrypted(encryptionCertificate);
         const bool isSigned(signingCertificate && signingKey);
-        std::uint32_t sigSize = 0;
+        std::uint16_t sigSize = 0;
 
         flatbuffers::Offset<flatbuffers::Vector<ByteVector::value_type>> flatbuffersBody;
         if (body.size() > 0) {
